+++ mitra_activitypub/src/importers.rs
@@ -750,7 +750,7 @@ async fn fetch_collection(
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct Collection {
-        id: CanonicalUri,
+        id: NonCanonicalUri,
         first: Option<JsonValue>, // page can be embedded
         #[serde(default)]
         items: Vec<JsonValue>,
@@ -760,7 +760,7 @@ async fn fetch_collection(
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct CollectionPage {
-        id: Option<CanonicalUri>,
+        id: Option<NonCanonicalUri>,
         next: Option<JsonValue>,
         #[serde(default)]
         items: Vec<JsonValue>,