+++ mitra_activitypub/src/builders/add_note.rs
@@ -11,6 +11,7 @@ use mitra_models::{
 use mitra_utils::id::generate_ulid;
 
 use crate::{
+    authority::Authority,
     contexts::{build_default_context, Context},
     deliverer::Recipient,
     identifiers::{
@@ -82,6 +83,7 @@ pub async fn prepare_add_note(
     sender: &User,
     post_id: Uuid,
 ) -> Result<OutgoingActivityJobData, DatabaseError> {
+    let authority = Authority::from(instance);
     let activity = build_add_note(
         instance.uri_str(),
         &sender.profile.username,
@@ -89,7 +91,7 @@ pub async fn prepare_add_note(
     );
     let recipients = get_add_note_recipients(db_client, sender.id).await?;
     Ok(OutgoingActivityJobData::new(
-        instance.uri_str(),
+        &authority,
         sender,
         activity,
         recipients,