+++ mitra_activitypub/src/builders/follow.rs
@@ -84,7 +84,7 @@ pub fn prepare_follow(
     );
     let recipients = Recipient::for_inbox(target_actor);
     Ok(OutgoingActivityJobData::new(
-        instance.uri_str(),
+        &authority,
         sender,
         activity,
         recipients,