+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Implemented `Deserialize` for `NonCanonicalUri` type.
+- Added `origin()` method to `NonCanonicalUri` type.
+
+### Removed
+
+- Removed incorrect implementation of `Deserialize` for `CanonicalUri`.
+
 ## [0.26.0] - 2026-07-02
 
 ### Added